requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.62.0

# Document parsing
PyPDF2>=2.10.0
pdfplumber>=0.7.0
python-docx>=0.8.11

# NLP libraries
transformers>=4.18.0
torch>=1.11.0
spacy>=3.2.0
nltk>=3.6.5

# API and Web
fastapi>=0.78.0
uvicorn>=0.17.6
pydantic>=1.9.0
python-multipart>=0.0.5

# Frontend
gradio>=3.0.0

# Testing and development
pytest>=7.0.0
black>=22.3.0
flake8>=4.0.1

# Docker and deployment
docker>=5.0.3